style: 데이터 페칭 네이밍 통일, 명시성 (httpclient 요청)
