+++ src/requests/place/fetchPlace.types.ts
@@ -1,0 +1,7 @@
+export interface FetchPlacesParams {
+  center: string;
+  radius?: number;
+  size?: number;
+  category?: string;
+  search?: string;
+}
