+++ src/utils/hooks/useAdminPlace/useAdminPlace.types.ts
@@ -1,4 +1,4 @@
-import { ResPlaceData } from '@/api/adminAPI/adminAPI.types';
+import { ResPlaceData } from '@/requests/admin/fetchAdmin.types';
 
 export interface UseGetAdminPlaceReturn {
   reportedPlaces: ResPlaceData[];
