+++ src/utils/hooks/useReview/useReview.types.ts
@@ -1,4 +1,4 @@
 import { UseQueryResult } from '@tanstack/react-query';
-import { ReviewResponse } from '@/api/reviewAPI/reviewAPI.types';
+import { ReviewResponse } from '@/requests/review/fetchReview.types';
 
 export type UseGetReviewsResult = UseQueryResult<ReviewResponse, Error>;
